fix(crosstab_count): accurately populate standard error = NA if count == 0
If the IPUMS sample doesn't have an observation of a given subgroup, then the mechanics of the standard error formula will make the standard errors look like they're 0. In actuality, we can't be confident that there truly are no individuals in that population who fall into that category. So instead, we override with standard_error = NA.